fix: grouped by function

+++ managment/src/main/java/com/example/managment/controllers/EmployeesController.java
@@ -1,14 +1,20 @@
 package com.example.managment.controllers;
 
 import com.example.managment.domain.employee.*;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
+@Validated
 @RestController
 @RequestMapping("/employees")
 public class EmployeesController {
@@ -18,14 +24,20 @@ public class EmployeesController {
     private EmployeeService employeeService;
     @GetMapping
     public ResponseEntity getAllEmployees(){
-        var allEmployees = repository.findAll();
+        try{
+            var allEmployees = repository.findAll();
 
-        return  ResponseEntity.ok(allEmployees);
+            return  ResponseEntity.ok(allEmployees);
+        }catch(Exception e){
+            throw new RuntimeException("Falha ao obter os dados");
+        }
+
     }
 
     @GetMapping("/ordem/alfabetica")
 
     public ResponseEntity getOrderName(){
+
         var employeesOrderByName = repository.findAllByOrderByNomeAsc();
 
         return  ResponseEntity.ok(employeesOrderByName);
@@ -62,10 +74,14 @@ public class EmployeesController {
     }
 
     @PostMapping
-    public ResponseEntity registerEmployee(@RequestBody RequestEmployee data){
-        Employee newEmployee = new Employee(data);
-        repository.save(newEmployee);
-        return ResponseEntity.ok().build();
+    public ResponseEntity registerEmployee(@Valid @RequestBody RequestEmployee data){
+        try{
+            Employee newEmployee = new Employee(data);
+            repository.save(newEmployee);
+            return ResponseEntity.ok().body("Novo funcionário cadastrado");
+        }catch(Exception e){
+            throw new EntityNotFoundException("Usuário não foi cadastrado");
+        }
     }
 
     @PostMapping("/aumento/salarial")
@@ -77,12 +93,13 @@ public class EmployeesController {
     @PutMapping ("/salario")
     public ResponseEntity getSalary(){
         var employeesSalary = repository.sumOfAllSalaries();
-        return ResponseEntity.ok(employeesSalary);
+
+        return ResponseEntity.ok().body("A soma dos salários é de: "+employeesSalary);
     }
 
     @PutMapping
     @Transactional
-    public ResponseEntity updateEmployee(@RequestBody RequestEmployee data){
+    public ResponseEntity updateEmployee(@Valid @RequestBody RequestEmployee data){
         Optional<Employee> optionalEmployee = repository.findById(data.id());
         if (optionalEmployee.isPresent()){
             Employee employee = optionalEmployee.get();
@@ -90,18 +107,23 @@ public class EmployeesController {
             employee.setData_nascimento(data.data_nascimento());
             employee.setSalario(data.salario());
             employee.setFuncao(data.funcao());
-            return ResponseEntity.ok(employee);
+            repository.save(employee);
+            return ResponseEntity.ok().body("Funcionário atualizado com sucesso");
         } else {
-            return ResponseEntity.notFound().build();
+            throw new EntityNotFoundException("Funcionario com esse id, não foi encontrado");
         }
     }
 
-
-
     @DeleteMapping("/{id}")
     public ResponseEntity deleteEmployee(@PathVariable String id){
-        Optional<Employee> optionalEmployee = repository.findById(id);
-        repository.deleteById(id);
-        return ResponseEntity.noContent().build();
+            Optional<Employee> optionalEmployee = repository.findById(id);
+            if (optionalEmployee.isPresent()){
+                repository.deleteById(id);
+                return ResponseEntity.ok().body("Funcionário com id: " + id + "deletado com sucesso");
+            }else {
+                throw new EntityNotFoundException();
+            }
     }
+
 }
+
